requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
greenlet==3.2.4
idna==3.10
lxml==6.0.0
playwright==1.54.0
pyee==13.0.0
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0

# RAG 챗봇을 위한 추가 패키지들
djangorestframework==3.15.2
django-cors-headers==4.4.0
chromadb==0.5.20
sentence-transformers==3.3.1
openai==1.58.1
python-dotenv==1.0.1
numpy==2.2.1
langchain
langchain-openai
langchain-chroma
